Move main App UI related components to separate directory (#36)
* Move main App UI related components to separate directory

* Convert AppError and initializing to ts
