+++ src/components/app/App.tsx
@@ -6,9 +6,9 @@ import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 
 import AppInitializing from './AppInitializing';
 import AppError from './AppError';
-import ChromInfoProvider from './ChromInfoProvider';
+import ChromInfoProvider from '../../ChromInfoProvider';
 import AppMain from './AppMain';
-import withEither from './with-either';
+import withEither from '../../with-either';
 
 import './App.css';
